refactor: update utility imports and clean up code structure
