+++ src/shared/components/ui/textarea.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/shared/lib/utils";
+import { cn } from "@/shared/utils/helpers";
 
 function Textarea({ className, ...props }: React.ComponentProps<"textarea">) {
   return (
